move analytics to own hook
