+++ applications/fishing-map/src/features/app/App.tsx
@@ -37,14 +37,14 @@ import { t } from 'features/i18n/i18n'
 import { useAppDispatch } from './app.hooks'
 import { selectAnalysisQuery, selectSidebarOpen } from './app.selectors'
 import styles from './App.module.css'
+import { useAnalytics } from './analytics.hooks'
+
 
 declare global {
   interface Window {
     gtag: any
   }
 }
-
-const GOOGLE_UNIVERSAL_ANALYTICS_ID = process.env.REACT_APP_GOOGLE_UNIVERSAL_ANALYTICS_ID
 
 const Main = () => {
   const workspaceLocation = useSelector(isWorkspaceLocation)
@@ -59,8 +59,8 @@ const Main = () => {
 }
 
 function App(): React.ReactElement {
+  useAnalytics()
   const dispatch = useAppDispatch()
-  const userData = useSelector(selectUserData)
   const sidebarOpen = useSelector(selectSidebarOpen)
   const { dispatchQueryParams } = useLocationConnect()
   const [menuOpen, setMenuOpen] = useState(false)
@@ -124,38 +124,6 @@ function App(): React.ReactElement {
   }, [])
 
   const { debugActive, dispatchToggleDebugMenu } = useDebugMenu()
-
-  useEffect(() => {
-    if (userData && GOOGLE_UNIVERSAL_ANALYTICS_ID && window.gtag) {
-      window.gtag('config', GOOGLE_UNIVERSAL_ANALYTICS_ID, {
-        user_id: userData.id,
-        custom_map: {
-          dimension1: 'userId',
-          dimension3: 'userGroup',
-          dimension4: 'userOrgType',
-          dimension5: 'userOrganization',
-          dimension6: 'userCountry',
-          dimension7: 'userLanguage',
-        },
-      })
-      window.gtag('set', 'user_properties', {
-        user_id: userData.id,
-        user_group: userData.groups,
-        user_org_type: userData.organizationType,
-        user_organization: userData.organization,
-        user_country: userData.country,
-        user_language: userData.language,
-      })
-      window.gtag('event', 'login', {
-        userId: userData.id,
-        userGroup: userData.groups,
-        userOrgType: userData.organizationType,
-        userOrganization: userData.organization,
-        userCountry: userData.country,
-        userLanguage: userData.language,
-      })
-    }
-  }, [userData])
 
   useEffect(() => {
     dispatch(fetchUserThunk())
